[Android] Use depfiles when generating projects and AARs
Contrary to what one would expect, having a dependency on a shared
library target such as `libxwalkcore` does not cause an action to be
re-run whenever the aforementioned shared library changes -- the .ninja
file only depends on the library's .TOC file, so the action is called
only if the library's symbol list changes.

In practice, this means rebuilding `libxwalkcore.so` would not cause
`xwalk_core_library` and its respective AAR to be regenerated, leading
to all sorts of problems.

We work around the issue by adding `depfile`s to the GN actions, and
then adding those native libraries (as well as some other files like the
.js ones in xwalk_core_library) to them. GN can then use their contents
and trigger the actions when the files there change (in addition to the
ones regularly tracked by the build system).

BUG=XWALK-7433

(cherry picked from commit 0556777edf280ef272a8b894122489cf67781bdd)