[Android] Do not require certificate transparency
Certificate transparency is a unique feature of Chrome but non-standard
SSL protocol. It aims to improve the security aloneside with CA. But
Crosswalk is based on content shell so the network connections initiated
by it do not support this certificate policy. This causes a SSL error
called ERT_STATUS_CERTIFICATE_TRANSPARENCY_REQUIRED and makes the app
continue to pop up the "SSL Certificate Error Alert" dialog.

This patch declares that Crosswalk doesn't require certificate
transparency to bypass the SSL error above.

BUG=XWALK-7398